pyproject.toml
[project]
name = "youtube-shorts-maker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.12.0",
    "google-genai>=1.31.0",
    "ipykernel>=6.30.1",
    "litellm>=1.76.0",
    "openai>=1.101.0",
]
